Divide classes into packages
Grouped classes into separate packages to improve code readability and maintainability
